requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
google-auth==2.43.0
google-genai==1.54.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.4
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
websockets==15.0.1
whitenoise
psycopg2-binary
django-filter
django-crispy-forms
crispy-bootstrap5
django-allauth
dj-database-url
psycopg2-binary
gunicorn
